backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
pydantic==2.10.5
pydantic-settings==2.7.1
python-multipart==0.0.20

# 数据获取
akshare==1.15.46
tushare==1.2.89  # Dual data source
pandas==2.2.3
numpy==1.26.4

# GPU加速 (DGX Spark)
cupy-cuda12x==13.3.0
numba==0.60.0
nvidia-ml-py==12.560.30

# RAPIDS (optional, for DGX Spark)
cudf-cu12==24.12.0  # GPU accelerated pandas

# 机器学习
scikit-learn==1.6.1
xgboost==2.1.3
lightgbm==4.5.0

# 缓存
redis==5.0.1
hiredis==2.3.2

# 定时任务
apscheduler==3.11.0

# WebSocket
websockets==14.2
python-socketio==5.12.1

# 数据库
sqlalchemy==2.0.37
aiosqlite==0.21.0

# 工具
python-dateutil==2.9.0
pytz==2024.2
ta-lib==0.6.0

# 监控
psutil==6.1.1
prometheus-client==0.21.1  # Production metrics

# 测试
pytest==8.3.5
pytest-asyncio==0.25.2
httpx==0.28.1
